Moved duplicated logic from both applications (Fulfillment and Usage) to common abstract application.
