+++ src/AutomationEngine.php
@@ -1,0 +1,65 @@
+<?php
+
+/**
+ * This file is part of the Ingram Micro Cloud Blue Connect SDK.
+ *
+ * @copyright (c) 2019. Ingram Micro. All Rights Reserved.
+ */
+
+namespace Connect;
+
+use Pimple\Container;
+use Pimple\Psr11\Container as PSRContainer;
+use Psr\Container\ContainerInterface;
+
+/**
+ * Class AutomationEngine
+ * @package Connect
+ */
+abstract class AutomationEngine
+{
+    /**
+     * Internal Dependency Container
+     * @var ContainerInterface
+     */
+    private $_container;
+
+    /**
+     * FulfillmentAutomation constructor.
+     * @param Config|null $configuration
+     * @param Container $container
+     * @throws ConfigException
+     */
+    public function __construct(Config $configuration = null, Container $container = null)
+    {
+        if (!isset($configuration)) {
+            $configuration = new Config('./config.json');
+        }
+
+        if (!isset($container)) {
+            $container = new Container();
+        }
+
+        $container['config'] = $configuration;
+
+        foreach ($configuration->runtimeServices as $id => $serviceProvider) {
+            if (!isset($container[$id]) && class_exists($serviceProvider, true)) {
+                $container[$id] = new $serviceProvider();
+            }
+        }
+
+        $this->_container = new PSRContainer($container);
+    }
+
+    /**
+     * Provide an access to the common libraries of the controller
+     * @param string $id
+     * @return object
+     */
+    public function __get($id)
+    {
+        return ($this->_container->has($id))
+            ? $this->_container->get($id)
+            : null;
+    }
+}
