+++ src/UsageAutomation.php
@@ -10,12 +10,9 @@ namespace Connect;
 
 use Connect\Usage\FileUsageRecord;
 use GuzzleHttp\ClientInterface;
-use Pimple\Container;
-use Pimple\Psr11\Container as PSRContainer;
-use Psr\Container\ContainerInterface;
-use Psr\Log\LoggerInterface;
 use PhpOffice\PhpSpreadsheet\Spreadsheet;
 use PhpOffice\PhpSpreadsheet\Writer\Xlsx;
+use Psr\Log\LoggerInterface;
 
 /**
  * Class UsageAutomation
@@ -25,52 +22,8 @@ use PhpOffice\PhpSpreadsheet\Writer\Xlsx;
  *
  * @package Connect
  */
-abstract class UsageAutomation implements UsageAutomationInterface
+abstract class UsageAutomation extends AutomationEngine implements UsageAutomationInterface
 {
-    /**
-     * Internal Dependency Container
-     * @var ContainerInterface
-     */
-    private $_container;
-
-    /**
-     * FulfillmentAutomation constructor.
-     * @param Config|null $configuration
-     * @param Container $container
-     * @throws ConfigException
-     */
-    public function __construct(Config $configuration = null, Container $container = null)
-    {
-        if (!isset($configuration)) {
-            $configuration = new Config('./config.json');
-        }
-
-        if (!isset($container)) {
-            $container = new Container();
-        }
-
-        $container['config'] = $configuration;
-
-        foreach ($configuration->runtimeServices as $id => $serviceProvider) {
-            if (!isset($container[$id]) && class_exists($serviceProvider, true)) {
-                $container[$id] = new $serviceProvider();
-            }
-        }
-
-        $this->_container = new PSRContainer($container);
-    }
-
-    /**
-     * Provide an access to the common libraries of the controller
-     * @param string $id
-     * @return object
-     */
-    public function __get($id)
-    {
-        return ($this->_container->has($id))
-            ? $this->_container->get($id)
-            : null;
-    }
 
     /**
      * Send the actual request to the connect endpoint
